✨ ERP：完成 product 信息、分类、单位的实现

+++ yudao-module-erp/yudao-module-erp-biz/src/main/java/cn/iocoder/yudao/module/erp/dal/dataobject/product/ErpProductDO.java
@@ -45,9 +45,9 @@ public class ErpProductDO extends BaseDO {
     /**
      * 单位编号
      *
-     * TODO 芋艿，关联
+     * 关联 {@link ErpProductUnitDO#getId()}
      */
-    private Integer unitId;
+    private Long unitId;
     /**
      * 产品状态
      *
